Split database test file
